Created common package

+++ data-validation-rest/src/main/java/com/aeris/datavalidationrest/instruments/InstrumentResource.java
@@ -1,5 +1,10 @@
 package com.aeris.datavalidationrest.instruments;
 
+import com.aeris.datavalidationrest.auth.LoginResource;
+import com.aeris.datavalidationrest.common.CommonService;
+import org.apache.http.HttpStatus;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -18,7 +23,9 @@ public class InstrumentResource {
     @Autowired
     private InstrumentDao instrumentDao;
     @Autowired
-    private InstrumentService instrumentService;
+    private CommonService commonService;
+
+    Logger logger = LoggerFactory.getLogger(LoginResource.class);
 
     @GetMapping
     public List<Instrument> findAll() {
@@ -34,34 +41,52 @@ public class InstrumentResource {
     }
 
     @PostMapping
-    public ResponseEntity<Void> add(@RequestBody @Valid Instrument instrument) {
-            //ADMIN
-            Instrument instrumentAdded;
+    public ResponseEntity<String> add(@RequestBody @Valid Instrument instrument) {
+        Instrument instrumentAdded;
 
-            if(instrument == null)
-                return ResponseEntity.noContent().build();
+        if(instrument == null)
+            return ResponseEntity.noContent().build();
 
-            this.instrumentService.getCurrentUserRole(request);
-
-            //instrumentAdded = instrumentDao.save(instrument);
+        if (this.commonService.isAdmin(request)) {
+            instrumentAdded = instrumentDao.save(instrument);
 
             URI location = ServletUriComponentsBuilder
                     .fromCurrentRequest()
                     .path("/{id}")
-                    //.buildAndExpand(instrumentAdded.getUuid())
-                    .buildAndExpand("Hello world !")
+                    .buildAndExpand(instrumentAdded.getUuid())
                     .toUri();
 
             return ResponseEntity.created(location).build();
+        }
+
+        return ResponseEntity.status(HttpStatus.SC_FORBIDDEN).body("Error Message");
     }
 
     @DeleteMapping(value = "/{id}")
-    public void delete(@PathVariable Instrument instrument) {//ADMIN and PI
-        instrumentDao.delete(instrument);
+    public  ResponseEntity<String> delete(@PathVariable Instrument instrument) {
+
+        if(instrument == null)
+            return ResponseEntity.noContent().build();
+
+        if (this.commonService.isAdmin(request)) {
+            instrumentDao.delete(instrument);
+            return ResponseEntity.status(HttpStatus.SC_OK).body("Update");
+        }
+
+        return ResponseEntity.status(HttpStatus.SC_FORBIDDEN).body("Error Message");
     }
 
     @PutMapping(value = "/update")
-    public void update(@RequestBody Instrument instrument) {
-        instrumentDao.save(instrument);
+    public ResponseEntity<String> update(@RequestBody Instrument instrument) {
+
+        if(instrument == null)
+            return ResponseEntity.noContent().build();
+
+        if (this.commonService.isAdmin(request)) {
+            instrumentDao.save(instrument);
+            return ResponseEntity.status(HttpStatus.SC_ACCEPTED).body("Update");
+        }
+
+        return ResponseEntity.status(HttpStatus.SC_FORBIDDEN).body("Error Message");
     }
 }
